Working #3139 - Catalogs default to host_config
The whole host_config concept is a bit outdated now that Configurer
exists, I think, (since any catalog it uses should be a host_config).
However, fixing that is outside of the scope of this series.

In the meantime, this does a better job of making sure every
catalog except the Settings catalogs are host_configs.

Signed-off-by: Luke Kanies <luke@puppetlabs.com>