Add binary package "puppet-testsuite"
Here is a patch to create a binary package (puppet-testsuite) that ships all
the tests available from the upstream source code. The test suites can then be
run by installing the package and running the tests directly from
/usr/share/puppet-testsuite/.

This was done at the request of the Ubuntu security team in order to improve
the QA process for the puppet package (in order to move it to main). The best
option would be to run it during the package build process - however it would
require more integration. Providing the tests as part of a binary packages is
the first step towards that goal.