[#3835] Fixed recursively absent directories improperly managing their files
Children of recursively absent (and only recursively absent) directories now
inherit the recursively absent behavior when they are created. This stops the
files from trying to be created, generating lots of  failure messages. This
doesn't affect directories which are absent and not recursive (whose children
aren't even attempted to be created, or directories which aren't absent.