Fix for pre 1.8.7 compatibility in namvar patch
There was a subtle 1.8.7 dependence in the composite key / namevar patch;
Nick discovered that our assumtion that hashes could be used as hash keys
does not hold in earlier bersions of ruby.  This patch replaces the hash
valued uniqueness_key with an array of the values of the (ordered) key
attributes.