Fix for #2807 Puppet settings available as variables
This is Luke's patch plus a change to fix a test that it broke.  It creates
a new sub-scope off the top scope, called "settings" and adds each of the
environment's settings to it as variables, thus satisfying the ticket while
taking us one step further from being able to implement futures.  *sigh*