Minimal fix for #4205 -- incorrect Import loop messages
This patch fixes the narrow problem of #4205, wherein type_loader would reparse
a file each time it was imported (causing the parser to incorrectly think that
it was in a loop) by checking @imported inside the existing check on the
thread-guarded @loaded.  (@imported was being set but never checked).

This works (and is thread safe) because all of this is going on inside a giant
synchronize care of @loaded.  But it, like the fix for #4208, does nothing
about the global lock.

Areas for future research:

1) Why is the looping inside of import?
2) Why are there separate @loaded and @imported tables?
3) Why is the parsing treated like a function (called deep in the structure)
   yet coded like a thread-savvy pseudo state monad (e.g. raising errors that
   presume it knows/owns what's going on outside the whole process)?

These and many other exciting questions are deferred to #4211