[#4213] -o option for setting onetime now works properly
When onetime was moved to global defaults, it broke the option handler
using it in agent to manage waitforcert length. Additionally, it
caused --onetime and -o to behave differently. This patch removes
the ordinary option handler defined in agent and moves the logic
for waitforcert to the one location it's used.