Minimal fix for #4243 -- import isn't thread safe
The import function was calling type_loader#import directly so that it could
pass in the current file name, but by doing so it was thwarting the thread-
safety locking level.  This patch rearanges things so that all imports go
through the same (thread safe) code path while retaining the current_file
passing, error handling, etc. from the old structure.