Feature: puppet-load - a tool to stress-test master compilation
This commit introduce a new executable (ext/puppet-load) which aims
to simulate concurrent clients to stress-test load a puppet master.
At the end of a run, it produces some statistics.

This tool is very lightweight:
 * it runs under Event Machine (and thus is event-driven)
 * it doesn't do anything with the received catalog

This tool, to run, needs access to:
 * a certificate/private_key pair from a node known by the master
 * a fact file like those persisted on the master
 * obviously a puppet-master running somewhere

Refer to the embedded help for options and run examples.

TODO:

 * fetch different nodes catalog
 * exercise the file server

Signed-off-by: Brice Figureau <brice-puppet@daysofwonder.com>