[#4344] Fix for failing templates when module name matches file in local dir.
When the name of a module matches the name of a file in the local
directory, puppet agent would sometimes try to read that file and
interpret it as puppet code.  This happened because files.rb was
unintentionally permitting puppet files without an extension.  Fixed
by changing the glob pattern to only permit ".pp" and ".rb"
extensions.