Moved perform_initial_import from Puppet::Resource::TypeCollection to Puppet::Node::Environment.
This change is part of an ongoing effort to remove functionality from
TypeCollection that is not related to keeping track of a collection of
types.  This reduces TypeCollection's linkage to the environment,
which is a step toward decoupling it from the type loading mechanism.

Also, added a spec test to verify that TypeCollection.version is
correctly recomputed when types are re-imported.