[#4685] Classes, defines, and nodes allowed inside of non-evaluated conditionals
Previously, ASTArray#evaluate() was responsible for checking whether
the user had tried to declare a class, define, or node in a prohibited
location (such as a conditional construct).  This meant that errors
would only be reported to the user if the conditional code was
actually evaluated.

Moved the checking into the parser, so that errors are always
reported.