[#4771] Import of manifests with the same name only happens once
The function import_if_possible, which was supposed to be responsible
for making sure that no two threads tried to import the same file at
the same time, was not making this decision based on the full pathname
of the file, since it was being invoked before pathnames were
resolved.  As a result, if we attempted to import two distinct files
with the same name at the same time (either in two threads or in a
single thread due to recursion), one of the files would not always get
imported.

Fixed this problem by moving the thread-safety logic to happen after
filenames are resolved to absolute paths.  This made it possible to
simplify the thread-safety logic significantly.