Fixes #4822 -- Puppet doc -o option broken
The global "-o" option ("--onetime") was overriding the
application-specific option "-o" because global options were being
sent to the OptionParser after application-specific options.

Modified the order in which options are sent to the OptionParser to
have the correct behavior.  Also merged together the two methods that
were applying options so that the order is more explicit.