Fixed #4025 (failure in launchd if certain plists are binary).
Modified the launchd provider to use OSX's "plutil" command to read
plists.  This allows it to handle properly lists in both XML format
and binary format.

Launchd continues to write out propertly lists in XML format.  This is
not a problem because the operating system is able to understand both
formats.