Fix for #4644: install.rb works properly on Windows
We were failing to make the sbin directory, and so the install script
was failing to install files to it. Also, we were trying to add man
pages regardless of whether or not we actually wanted to. Changed
operating system detection to use Facter["operatingsystem"], which is
easier to check than RUBY_PLATFORM (there are multiple different values
for Windows).