Fix for #4649 -- avoid unsupported features on non-posix systems
MS Windows is apparently only "posix curious;" many of the usual fcntl flags
are not defined when running under MS Windows.  It also appears that we are
only using a small subset of the vendored event-loop library and could replace
it with a handful of (portable) extracted methods.  This, however, would be
too agressive of a refactor for 2.6.2 and so this commit simply avoids referring
to the unsupported flag values in environments where they are not defined.