Fix for #4945 -- explicitly check os to supress man page installation
The fix for #4644 usurped an existing option flag to supress the installation
of man pages on Microsoft Windows, when the desired behaviour was to only skip
the installation on MSWin and not change the behaviour on other platforms.
This patch implements the check explicitly to correctly express the desired
behaviour.