(#4534/#4778) -- Normalize parameterized classes
This is a reconciliation/melding of Paul's
    (#4534) Class inheritance with parameterized classes is no longer ignored
and Markus's
    Fix for #4778 -- evaluate parameterized classes when they are instantiated

Extracted the code from Resource::Type#mk_plain_resource that evaluates
parents and tags the catalog, and moved that into a new method called
instantiate_resource.  Instantiate_resource is now also called from
Parser::Ast::Resource#evaluate, so that the notation

    "class { classname: }"

now executes this code too.  Likewise adds class evaluation so that it behaves
the same (with regard to lazy / strict evaluation) as

    include classname