Fix #5020 - Prefer finding node name from REST uri over certname
This is a behavior change. Before this patch, we always used the currently
connected node's certname to compile the catalog, despite the value of
the catalog URI REST request.
With this patch we now use the URI as the compiled node name.

This is safe because the default auth.conf (and default inserted rules
when no auth.conf is present) only allow the given connected node to
compile its own catalog.

But this also allows for greater flexibility with auth.conf. For instance
it can be used by a monitoring system to check multiple nodes catalogs
with only one certificate:

path ~ ^/catalog/([^/]+)$
method find
allow $1
allow monitoring-station.domain.com

Signed-off-by: Brice Figureau <brice-puppet@daysofwonder.com>