Fix #5023 - puppet-load multiple nodes support
This patch allows puppet-load to compile multiple nodes catalog.
This is done by using multiple --node. Puppet-load will use round-robin
to chose which nodes catalog to ask for a given simulated client.
It is also possible to pass a directory of facts yaml file in which
puppet-load will load given --node facts file.

This can work only if #5020 is applied to the puppetmaster first.

Signed-off-by: Brice Figureau <brice-puppet@daysofwonder.com>